feat: add MonsterHeroSystems in pico_ecs
* fix entt system with group

+++ src/base/base/components/VelocityComponent.h
@@ -4,8 +4,8 @@
 namespace ecs::benchmarks::base::components {
 
 struct VelocityComponent {
-  float x{1.0F};
-  float y{1.0F};
+  double x{1.0};
+  double y{1.0};
 };
 
 } // namespace ecs::benchmarks::base::components
